Fix for gcc-3.4
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@9267 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/net/multicast.cxx
@@ -246,7 +246,7 @@ int			openMulticast(const Address& address, int port,
   /* check interface */
   if (net_interface && net_interface[0]) {
     ifaddr.s_addr = inet_addr(net_interface);
-    if (ifaddr.s_addr == (unsigned long)-1) {
+    if (ifaddr.s_addr == (uint32_t)-1) {
       struct hostent* hp = gethostbyname(net_interface);
       if (!hp) {
 	std::vector<std::string> args;
